Fleshed out the GraphQL Server and Fixed some misaligned types

+++ alliance.go
@@ -7,13 +7,13 @@ import (
 )
 
 type Alliance struct {
-	ID                    uint64    `db:"id" json:"id"`
+	ID                    uint32    `db:"id" json:"id"`
 	Name                  string    `db:"name" json:"name"`
 	Ticker                string    `db:"ticker" json:"ticker"`
-	CreatorCorporationID  uint64    `db:"creator_corporation_id" json:"creator_corporation_id"`
+	CreatorCorporationID  uint32    `db:"creator_corporation_id" json:"creator_corporation_id"`
 	CreatorID             uint64    `db:"creator_id" json:"creator_id"`
 	DateFounded           null.Time `db:"date_founded" json:"date_founded"`
-	ExecutorCorporationID uint64    `db:"executor_corporation_id" json:"executor_corporation_id"`
+	ExecutorCorporationID uint32    `db:"executor_corporation_id" json:"executor_corporation_id"`
 	Ignored               bool      `db:"ignored" json:"ignored"`
 	Closed                bool      `db:"closed" json:"closed"`
 	Expires               time.Time `db:"expires" json:"expires"`
